Move TimeStackedBarChart to components/index/cards/charts
